[io-sys] move solve_lyap/solve_ricc implementations to bindings package
