+++ src/pymor/algorithms/riccati.py
@@ -2,159 +2,18 @@
 # Copyright 2013-2016 pyMOR developers and contributors. All rights reserved.
 # License: BSD 2-Clause License (http://opensource.org/licenses/BSD-2-Clause)
 
-from __future__ import absolute_import, division, print_function
-
-import numpy as np
-
-from pymor.algorithms.to_matrix import to_matrix
-from pymor.operators.interfaces import OperatorInterface
-from pymor.operators.constructions import IdentityOperator, LincombOperator
+from pymor.core.config import config
+from pymor.core.defaults import defaults
 
 
-def operator2matrix(A):
-    """Transforms NumpyMatrixOperator to NumPy array
-
-    Parameters
-    ----------
-    A
-        NumpyMatrixOperator or `None`.
-
-    Returns
-    -------
-    A_mat
-        |NumPy array| or `None`.
-    """
-    if A is None:
-        A_mat = None
-    else:
-        A_mat = to_matrix(A)
-    return A_mat
+_DEFAULT_SOLVER = 'slycot' if config.HAVE_SLYCOT else \
+                  'pymess' if config.HAVE_PYMESS else \
+                  'scipy'
 
 
-try:
-    import pymess
-
-    class RiccatiEquation(pymess.equation):
-        r"""Riccati equation class for pymess
-
-        Represents a Riccati equation
-
-        .. math::
-            A^T X + X A - X B B^T X + C^T C = 0
-
-        if E is `None`, otherwise a generalized Lyapunov equation
-
-        .. math::
-            A^T X E + E^T X A - E^T X B B^T X E + C^T C = 0.
-
-        For the dual Riccati equation
-
-        .. math::
-            A X + X A^T - X C^T C X + B B^T = 0, \\
-            A X E^T + E X A^T - E X C^T C X E^T + B B^T = 0,
-
-        `opt.type` needs to be `pymess.MESS_OP_NONE`.
-
-        Parameters
-        ----------
-        opt
-            pymess options structure.
-        A
-            The |Operator| A.
-        E
-            The |Operator| E or `None`.
-        B
-            The |Operator| B.
-        C
-            The |Operator| C.
-        """
-        def __init__(self, opt, A, E, B, C):
-            super().__init__(name='ricc_eqn', opt=opt, dim=A.source.dim)
-
-            self.A = A
-            self.E = E
-            self.B = to_matrix(B)
-            self.C = to_matrix(C)
-            self.RHS = self.B if opt.type == pymess.MESS_OP_NONE else self.C.T
-            self.p = []
-
-        def AX_apply(self, op, y):
-            y = self.A.source.from_data(np.array(y).T)
-            if op == pymess.MESS_OP_NONE:
-                x = self.A.apply(y)
-            else:
-                x = self.A.apply_transpose(y)
-            return np.matrix(x.data).T
-
-        def EX_apply(self, op, y):
-            if self.E is None:
-                return y
-
-            y = self.A.source.from_data(np.array(y).T)
-            if op == pymess.MESS_OP_NONE:
-                x = self.E.apply(y)
-            else:
-                x = self.E.apply_transpose(y)
-            return np.matrix(x.data).T
-
-        def AINV_apply(self, op, y):
-            y = self.A.source.from_data(np.array(y).T)
-            if op == pymess.MESS_OP_NONE:
-                x = self.A.apply_inverse(y)
-            else:
-                x = self.A.apply_inverse_transpose(y)
-            return np.matrix(x.data).T
-
-        def EINV_apply(self, op, y):
-            if self.E is None:
-                return y
-
-            y = self.A.source.from_data(np.array(y).T)
-            if op == pymess.MESS_OP_NONE:
-                x = self.E.apply_inverse(y)
-            else:
-                x = self.E.apply_inverse_transpose(y)
-            return np.matrix(x.data).T
-
-        def ApEX_apply(self, op, p, idx_p, y):
-            y = self.A.source.from_data(np.array(y).T)
-            if op == pymess.MESS_OP_NONE:
-                x = self.A.apply(y)
-                if self.E is None:
-                    x += p * y
-                else:
-                    x += p * self.E.apply(y)
-            else:
-                x = self.A.apply_transpose(y)
-                if self.E is None:
-                    x += p.conjugate() * y
-                else:
-                    x += p.conjugate() * self.E.apply_transpose(y)
-            return np.matrix(x.data).T
-
-        def ApEINV_apply(self, op, p, idx_p, y):
-            y = self.A.source.from_data(np.array(y).T)
-            E = IdentityOperator(self.A.source) if self.E is None else self.E
-
-            if p.imag == 0:
-                ApE = LincombOperator((self.A, E), (1, p.real))
-            else:
-                ApE = LincombOperator((self.A, E), (1, p))
-
-            if op == pymess.MESS_OP_NONE:
-                x = ApE.apply_inverse(y)
-            else:
-                x = ApE.apply_inverse_transpose(y)
-            return np.matrix(x.data).T
-
-        def parameter(self, arp_p, arp_m, B=None, K=None):
-            return None
-except ImportError:
-    pass
-
-
+@defaults('me_solver')
 def solve_ricc(A, E=None, B=None, Q=None, C=None, R=None, G=None,
-               trans=False, me_solver=None, tol=None):
+               trans=False, me_solver=_DEFAULT_SOLVER, tol=None):
     """Find a factor of the solution of a Riccati equation
 
     Returns factor :math:`Z` such that :math:`Z Z^T` is approximately the
@@ -199,8 +58,7 @@ def solve_ricc(A, E=None, B=None, Q=None, C=None, R=None, G=None,
     trans
         If the dual equation needs to be solved.
     me_solver
-        Method to use ('scipy', 'slycot', 'pymess_care', 'pymess_lrnm').
-        If `me_solver` is `None`, a solver is chosen automatically.
+        Method to use ('scipy', 'slycot', 'pymess', 'pymess_care', 'pymess_lrnm').
     tol
         Tolerance parameter.
 
@@ -210,192 +68,15 @@ def solve_ricc(A, E=None, B=None, Q=None, C=None, R=None, G=None,
         Low-rank factor of the Riccati equation solution,
         |VectorArray| from `A.source`.
     """
-    assert isinstance(A, OperatorInterface) and A.linear
-    assert A.source == A.range
-    if E is not None:
-        assert isinstance(E, OperatorInterface) and E.linear
-        assert E.source == E.range == A.source
-    if not trans:
-        if C is not None:
-            assert Q is None
-            assert isinstance(C, OperatorInterface) and C.linear
-            assert C.source == A.source
-        else:
-            assert isinstance(Q, OperatorInterface) and Q.linear
-            assert Q.source == Q.range == A.source
-        if G is not None:
-            assert B is None and R is None
-            assert isinstance(G, OperatorInterface) and G.linear
-            assert G.source == G.range == A.source
-        else:
-            assert isinstance(B, OperatorInterface) and B.linear
-            assert B.range == A.source
-            if R is not None:
-                assert isinstance(R, OperatorInterface) and R.linear
-                assert R.source == R.range == B.source
-    else:
-        if B is not None:
-            assert Q is None
-            assert isinstance(B, OperatorInterface) and B.linear
-            assert B.range == A.source
-        else:
-            assert isinstance(Q, OperatorInterface) and Q.linear
-            assert Q.source == Q.range == A.source
-        if G is not None:
-            assert C is None and R is None
-            assert isinstance(G, OperatorInterface) and G.linear
-            assert G.source == G.range == A.source
-        else:
-            assert C is not None
-            assert isinstance(C, OperatorInterface) and C.linear
-            assert C.source == A.source
-            if R is not None:
-                assert isinstance(R, OperatorInterface) and R.linear
-                assert R.source == R.range == C.range
-    assert me_solver is None or me_solver in {'scipy', 'slycot', 'pymess_care', 'pymess_lrnm'}
-
-    if me_solver is None:
-        import imp
-        try:
-            imp.find_module('slycot')
-            me_solver = 'slycot'
-        except ImportError:
-            try:
-                imp.find_module('pymess')
-                if A.source.dim >= 1000:
-                    me_solver = 'pymess_lrnm'
-                else:
-                    me_solver = 'pymess_care'
-            except ImportError:
-                me_solver = 'scipy'
+    assert me_solver in {'scipy', 'slycot', 'pymess', 'pymess_care', 'pymess_lrnm'}
 
     if me_solver == 'scipy':
-        if E is not None or G is not None:
-            raise NotImplementedError()
-        import scipy.linalg as spla
-        A_mat = operator2matrix(A)
-        B_mat = operator2matrix(B)
-        C_mat = operator2matrix(C)
-        Q_mat = operator2matrix(Q)
-        R_mat = operator2matrix(R)
-        if R is None:
-            if not trans:
-                R_mat = np.eye(B.source.dim)
-            else:
-                R_mat = np.eye(C.range.dim)
-        if not trans:
-            if Q is None:
-                Q_mat = C_mat.T.dot(C_mat)
-            X = spla.solve_continuous_are(A_mat, B_mat, Q_mat, R_mat)
-        else:
-            if Q is None:
-                Q_mat = B_mat.dot(B_mat.T)
-            X = spla.solve_continuous_are(A_mat.T, C_mat.T, Q_mat, R_mat)
-        from pymor.algorithms.cholp import cholp
-        Z = cholp(X, copy=False)
+        from pymor.bindings.scipy import solve_ricc_impl
     elif me_solver == 'slycot':
-        import slycot
-        A_mat = operator2matrix(A)
-        B_mat = operator2matrix(B)
-        C_mat = operator2matrix(C)
-        R_mat = operator2matrix(R)
-        G_mat = operator2matrix(G)
-        Q_mat = operator2matrix(Q)
+        from pymor.bindings.slycot import solve_ricc as solve_ricc_impl
+    elif me_solver.startswith('pymess'):
+        from pymor.bindings.pymess import solve_ricc as solve_ricc_impl
+    else:
+        assert False
 
-        n = A_mat.shape[0]
-        dico = 'C'
-
-        if E is None:
-            if not trans:
-                if G is None:
-                    if R is None:
-                        G_mat = B_mat.dot(B_mat.T)
-                    else:
-                        G_mat = slycot.sb02mt(n, B_mat.shape[1], B_mat, R_mat)[-1]
-                if C is not None:
-                    Q_mat = C_mat.T.dot(C_mat)
-                X = slycot.sb02md(n, A_mat, G_mat, Q_mat, dico)[0]
-            else:
-                if G is None:
-                    if R is None:
-                        G_mat = C_mat.T.dot(C_mat)
-                    else:
-                        G_mat = slycot.sb02mt(n, C_mat.shape[0], C_mat.T, R_mat)[-1]
-                if B is not None:
-                    Q_mat = B_mat.dot(B_mat.T)
-                X = slycot.sb02md(n, A_mat.T, G_mat, Q_mat, dico)[0]
-        else:
-            E_mat = operator2matrix(E)
-            jobb = 'B' if G is None else 'B'
-            fact = 'C' if Q is None else 'N'
-            uplo = 'U'
-            jobl = 'Z'
-            scal = 'N'
-            sort = 'S'
-            acc = 'R'
-            if not trans:
-                m = 0 if B is None else B_mat.shape[1]
-                p = 0 if C is None else C_mat.shape[0]
-                if G is not None:
-                    B_mat = G_mat
-                    R_mat = np.empty((1, 1))
-                elif R is None:
-                    R_mat = np.eye(m)
-                if Q is None:
-                    Q_mat = C_mat
-                L_mat = np.empty((n, m))
-                ret = slycot.sg02ad(dico, jobb, fact, uplo, jobl, scal, sort, acc, n, m, p,
-                                    A_mat, E_mat, B_mat, Q_mat, R_mat, L_mat)
-            else:
-                m = 0 if C is None else C_mat.shape[0]
-                p = 0 if B is None else B_mat.shape[1]
-                if G is not None:
-                    C_mat = G_mat
-                    R_mat = np.empty((1, 1))
-                elif R is None:
-                    C_mat = C_mat.T
-                    R_mat = np.eye(m)
-                if Q is None:
-                    Q_mat = B_mat.T
-                L_mat = np.empty((n, m))
-                ret = slycot.sg02ad(dico, jobb, fact, uplo, jobl, scal, sort, acc, n, m, p,
-                                    A_mat.T, E_mat.T, C_mat, Q_mat, R_mat, L_mat)
-            X = ret[1]
-            iwarn = ret[-1]
-            if iwarn == 1:
-                print('slycot.sg02ad warning: solution may be inaccurate.')
-        from pymor.algorithms.cholp import cholp
-        Z = cholp(X, copy=False)
-    elif me_solver == 'pymess_care':
-        if Q is not None or R is not None or G is not None:
-            raise NotImplementedError()
-        import pymess
-        A_mat = operator2matrix(A)
-        E_mat = operator2matrix(E)
-        B_mat = operator2matrix(B)
-        C_mat = operator2matrix(C)
-        if not trans:
-            Z = pymess.care(A_mat, E_mat, B_mat, C_mat)
-        else:
-            if E is None:
-                Z = pymess.care(A_mat.T, None, C_mat.T, B_mat.T)
-            else:
-                Z = pymess.care(A_mat.T, E_mat.T, C_mat.T, B_mat.T)
-    elif me_solver == 'pymess_lrnm':
-        if Q is not None or R is not None or G is not None:
-            raise NotImplementedError()
-        import pymess
-        opts = pymess.options()
-        opts.adi.shifts.paratype = pymess.MESS_LRCFADI_PARA_ADAPTIVE_V
-        if not trans:
-            opts.type = pymess.MESS_OP_TRANSPOSE
-        if tol is not None:
-            opts.rel_change_tol = tol
-            opts.adi.res2_tol = tol
-            opts.adi.res2c_tol = tol
-        eqn = RiccatiEquation(opts, A, E, B, C)
-        Z, status = pymess.lrnm(eqn, opts)
-
-    Z = A.source.from_data(np.array(Z).T)
-
-    return Z
+    return solve_ricc_impl(A, E=E, B=B, Q=Q, C=C, R=R, G=G, trans=trans, me_solver=me_solver, tol=tol)
